Use same hook to resolve democracy proposals and referendums
