+++ src/api/hooks/useDemocracy.ts
@@ -1,15 +1,109 @@
-import useApiQuery from 'src/api/hooks/useApiQuery';
+import {gql, useQuery, NetworkStatus} from '@apollo/client';
+import type {
+  SubstrateChainDemocracyProposal,
+  SubstrateChainDemocracyReferendum,
+} from 'src/generated/litentryGraphQLTypes';
+
+export type DemocracyProposal = SubstrateChainDemocracyProposal;
+export type DemocracyReferendum = SubstrateChainDemocracyReferendum;
+
+type Democracy = {
+  substrateChainDemocracyProposals: DemocracyProposal[];
+  substrateChainDemocracyReferendums: DemocracyReferendum[];
+};
+
+export const ACCOUNT_FIELDS = gql`
+  fragment AccountFields on SubstrateChainAccount {
+    address
+    display
+    registration {
+      display
+      displayParent
+      email
+      image
+      legal
+      pgp
+      riot
+      twitter
+      web
+      judgements {
+        index
+        judgement {
+          isUnknown
+          isFeePaid
+          isReasonable
+          isKnownGood
+          isOutOfDate
+          isLowQuality
+          isErroneous
+        }
+      }
+    }
+  }
+`;
+
+export const DEMOCRACY_QUERY = gql`
+  ${ACCOUNT_FIELDS}
+  query getDemocracy {
+    substrateChainDemocracyProposals {
+      index
+      balance
+      formattedBalance
+      meta
+      method
+      section
+      hash
+      proposer {
+        address
+        account {
+          ...AccountFields
+        }
+      }
+      seconds {
+        address
+        account {
+          ...AccountFields
+        }
+      }
+      args {
+        name
+        type
+        value
+      }
+    }
+    substrateChainDemocracyReferendums {
+      index
+      meta
+      method
+      section
+      hash
+      endPeriod
+      activatePeriod
+      votedAye
+      formattedVotedAye
+      votedNay
+      formattedVotedNay
+      voteCountAye
+      voteCountNay
+      ayePercent
+      args {
+        name
+        type
+        value
+      }
+    }
+  }
+`;
 
 export function useDemocracy() {
-  return useApiQuery(['democracy'], async (api) => {
-    const [referendums, activeProposals] = await Promise.all([
-      api.derive.democracy.referendums(),
-      api.derive.democracy.proposals(),
-    ]);
+  const {data, networkStatus, ...rest} = useQuery<Democracy>(DEMOCRACY_QUERY);
 
-    return {
-      activeProposals,
-      referendums,
-    };
-  });
+  return {
+    data: {
+      proposals: data?.substrateChainDemocracyProposals,
+      referendums: data?.substrateChainDemocracyReferendums,
+    },
+    refreshing: networkStatus === NetworkStatus.refetch,
+    ...rest,
+  };
 }
